ændret ints til Longs ifbm. Repos.
 ResponseEntity tilføjet i Controller.

 SQL Workbench forbundet på min (Patrick)'s pc til IntelliJ og forsøger at løse problemer med SQL syntaks, da der er uoverensstemmelser mellem database navne og i IntelliJ entiteter.'

 Forsøger til sidst også at få TaxiTest til at køre, når SQL syntaks virker.

+++ src/main/java/group_3/airport_admin_system/model/Gate.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
 public class Gate {
 
     @Id
-    private int gateNumber;
+    private Long gateNumber;
 
     @Column(nullable = false, length = 1)
     private Wakecategory wake;
@@ -17,7 +17,7 @@ public class Gate {
 
     public Gate(){ }
 
-    public int getGateNumber() {
+    public Long getGateNumber() {
         return gateNumber;
     }
